+++ Project_4/Project_4/numerical2.h
@@ -3,4 +3,4 @@
 #include <mpi.h>
 using namespace arma;
 using namespace std;
-void Numerical2(int T, double beta, int L, bool random, int my_rank, mat &local_expectation_values);
+void Numerical2(long int T, double beta, int L, bool random, int my_rank, mat &local_expectation_values);
